add edit landscape (mac pro commit)

+++ app/src/main/java/ru/kitfactory/mixingautopaint/data/storage/db/Paint.kt
@@ -12,9 +12,9 @@ data class Paint(
     @PrimaryKey(autoGenerate = true)
     val id: Int, //уникальное поле
     var titleMix: String = "", // заголовок
-    var partPaint: Int = 0, //количество частей краски
-    var partHardener: Int = 0, // количество частей отвердителя
-    var partDiluent: Int = 0, // колличество частей разбавителя
+    var partPaint: Float = 0f, //количество частей краски
+    var partHardener: Float = 0f, // количество частей отвердителя
+    var partDiluent: Float = 0f, // колличество частей разбавителя
     var paintMass: Int = 0, // масса краски
     var paintMassForMix: Float = 0f,
     var massHardenerForMix: Float = 0f, // масса отвердителя
