+++ app/src/main/java/ru/kitfactory/mixingautopaint/presentation/addFragment/AddMixPaintFragment.kt
@@ -79,9 +79,9 @@ class AddMixPaintFragment : Fragment() {
             val paint = Paint(
                 0,
                 forMix.title,
-                forMix.paintPart.toInt(),
-                forMix.hardenerPart.toInt(),
-                forMix.diluentPart.toInt(),
+                forMix.paintPart.toFloat(),
+                forMix.hardenerPart.toFloat(),
+                forMix.diluentPart.toFloat(),
                 forMix.massPaint.toInt(),
                 mixPaint.massPaint,
                 mixPaint.massHardener,
